+++ src/effect/base.ts
@@ -6,15 +6,6 @@ import BaseObject from '../object'
  * @deprecated All visual effects now inherit from `Visual` instead
  */
 export class Base implements BaseObject {
-  static readonly Event = {
-    /**
-     * Fired when the effect is ready to be applied
-     *
-     * @event
-     */
-    READY: 'ready'
-  }
-
   type: string
   publicExcludes: string[]
   propertyFilters: Record<string, <T>(value: T) => T>
@@ -34,6 +25,11 @@ export class Base implements BaseObject {
     this._occurrenceCount = 0
     this._target = null
   }
+
+  /**
+   * Wait until this effect is ready to be applied
+   */
+  async whenReady (): Promise<void> {} // eslint-disable-line @typescript-eslint/no-empty-function
 
   /**
    * Attaches this effect to `target` if not already attached.