+++ src/layer/image.ts
@@ -1,4 +1,3 @@
-import { publish } from '../event'
 import { VisualSource, VisualSourceOptions } from './visual-source'
 
 type ImageOptions = VisualSourceOptions
@@ -11,13 +10,6 @@ class Image extends VisualSource {
       options.source = img
     }
     super(options)
-
-    // Emit ready event when the image is ready to be drawn
-    this.source.addEventListener('load', () => {
-      // Make sure all superclasses are ready
-      if (this.ready)
-        publish(this, Image.Event.READY, {})
-    })
   }
 }